+++ data-manager/api/src/test/java/org/orbisgis/orbisdata/datamanager/api/dataset/ITableTest.java
@@ -564,6 +564,12 @@ public class ITableTest {
             return null;
         }
 
+        @NotNull
+        @Override
+        public Map<String, Object> firstRow() {
+            return null;
+        }
+
         @Override
         public String getLocation() {
             return null;