+++ data-manager/api/src/test/java/org/orbisgis/orbisdata/datamanager/api/datasource/IJdbcDataSourceTest.java
@@ -499,6 +499,12 @@ public class IJdbcDataSourceTest {
             return false;
         }
 
+        @Nullable
+        @Override
+        public Collection<String> getColumnNames(String location) {
+            return null;
+        }
+
         @Override
         public Connection getConnection() throws SQLException {
             return null;