+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/postgis/POSTGIS.java
@@ -21,6 +21,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.sql.Connection;
+import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.HashMap;
@@ -150,7 +151,15 @@ public class POSTGIS extends JdbcDataSource {
         }
         StatementWrapper statement;
         try {
-            statement = (StatementWrapper) connectionWrapper.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_UPDATABLE);
+            DatabaseMetaData dbdm = connectionWrapper.getMetaData();
+            int type = ResultSet.TYPE_FORWARD_ONLY;
+            if(dbdm.supportsResultSetType(ResultSet.TYPE_SCROLL_SENSITIVE)){
+                type = ResultSet.TYPE_SCROLL_SENSITIVE;
+            }
+            else if(dbdm.supportsResultSetType(ResultSet.TYPE_SCROLL_INSENSITIVE)){
+                type = ResultSet.TYPE_SCROLL_INSENSITIVE;
+            }
+            statement = (StatementWrapper) connectionWrapper.createStatement(type, ResultSet.CONCUR_UPDATABLE);
         } catch (SQLException e) {
             LOGGER.error("Unable to create Statement.\n" + e.getLocalizedMessage());
             return null;
@@ -182,7 +191,15 @@ public class POSTGIS extends JdbcDataSource {
         }
         StatementWrapper statement;
         try {
-            statement = (StatementWrapper) connectionWrapper.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_UPDATABLE);
+            DatabaseMetaData dbdm = connectionWrapper.getMetaData();
+            int type = ResultSet.TYPE_FORWARD_ONLY;
+            if(dbdm.supportsResultSetType(ResultSet.TYPE_SCROLL_SENSITIVE)){
+                type = ResultSet.TYPE_SCROLL_SENSITIVE;
+            }
+            else if(dbdm.supportsResultSetType(ResultSet.TYPE_SCROLL_INSENSITIVE)){
+                type = ResultSet.TYPE_SCROLL_INSENSITIVE;
+            }
+            statement = (StatementWrapper) connectionWrapper.createStatement(type, ResultSet.CONCUR_UPDATABLE);
         } catch (SQLException e) {
             LOGGER.error("Unable to create Statement.\n" + e.getLocalizedMessage());
             return null;